Renaming the stickyScroll to stickyScrollController
