+++ src/vs/editor/contrib/stickyScroll/browser/stickyScrollContribution.ts
@@ -5,7 +5,7 @@
 
 import { registerEditorContribution } from 'vs/editor/browser/editorExtensions';
 import { ToggleStickyScroll } from 'vs/editor/contrib/stickyScroll/browser/stickyScrollActions';
-import { StickyScrollController } from 'vs/editor/contrib/stickyScroll/browser/stickyScroll';
+import { StickyScrollController } from 'vs/editor/contrib/stickyScroll/browser/stickyScrollController';
 import { registerAction2 } from 'vs/platform/actions/common/actions';
 
 registerEditorContribution(StickyScrollController.ID, StickyScrollController);
