+++ src/vs/editor/contrib/stickyScroll/browser/stickyScrollController.ts
